requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
openpyxl==3.1.2
qrcode==7.4.2
pillow==10.1.0 